Modify Function Interface
