+++ Project.toml
@@ -1,7 +1,7 @@
 name    = "PseudoBooleanOptimization"
 uuid    = "c8fa9a04-bc42-452d-8558-dc51757be744"
 authors = ["pedromxavier <pedrox@cos.ufrj.br>"]
-version = "0.1.1"
+version = "0.2.0"
 
 [deps]
 LinearAlgebra      = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -10,4 +10,4 @@ Random             = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 
 [compat]
 MutableArithmetics = "1.3"
-julia = "1.6"
+julia              = "1.6"
