+++ benchmark/suites/constructors.jl
@@ -1,4 +1,4 @@
-function benchmark_constructors!(suite, ::Type{F}) where {V,T,F<:PBO.AbstractFunction{V,T}}
+function benchmark_constructors!(suite, ::Type{F}) where {V,T,F<:PBO.AbstractPBF{V,T}}
     suite["constructors"] = BenchmarkGroup()
 
     return nothing
