+++ src/library/function/function.jl
@@ -1,0 +1,7 @@
+@doc raw"""
+    PseudoBooleanFunction{V,T,X}
+"""
+struct PseudoBooleanFunction{V,T,S} <: AbstractPseudoBooleanFunction{V,T}
+    Ω::S
+end
+
