+++ src/library/synthesis/not_all_equal_3sat.jl
@@ -5,7 +5,7 @@ Generates Not-all-equal 3-SAT problem with ``m`` variables and ``n`` clauses.
 """
 function not_all_equal_3sat end
 
-function not_all_equal_3sat(rng, ::Type{F}, n::Integer, m::Integer) where {V,T,F<:AbstractFunction{V,T}}
+function not_all_equal_3sat(rng, ::Type{F}, n::Integer, m::Integer) where {V,T,F<:AbstractPBF{V,T}}
     J = Dict{Tuple{Int,Int},T}()
 
     C = BitSet(1:n)
@@ -31,9 +31,12 @@ function not_all_equal_3sat(rng, ::Type{F}, n::Integer, m::Integer) where {V,T,F
 
     # Convert to boolean
     # s = 2x - 1
-    return sum([J[i,j] * F(i => T(2), T(-1)) * F(j => T(2), T(-1)) for (i,j) in keys(J)])
+    f = sum([J[i,j] * F(i => T(2), T(-1)) * F(j => T(2), T(-1)) for (i,j) in keys(J)])
+    x = nothing # no planted solutions
+
+    return (f, x)
 end
 
-function not_all_equal_3sat(::Type{F}, n::Integer, m::Integer) where {V,T,F<:AbstractFunction{V,T}}
+function not_all_equal_3sat(::Type{F}, n::Integer, m::Integer) where {V,T,F<:AbstractPBF{V,T}}
     return not_all_equal_3sat(Random.GLOBAL_RNG, F, n, m)
 end
